Rename fns and vars to be more accurate and descriptive
The original "useCryptoText" wasn't actually a hook to begin with. "useCryptoTextSimple" was the real hook.

- Rename/move useCryptoText->getCryptoText in a separate cryptoTextUtils.ts
- Rename/move useCryptoTextSimple->useCryptoText, move to useCryptoText.ts
